chore: Use @guardian/cdk mock for tracking tag
The value of the tracking tag maps to the version of @guardian/cdk used.

This means bumping the version in package.json isn't merely enough as the snapshots will be out of date.
That is, PRs raised by Dependabot to automatically update @guardian/cdk will always fail.

By using the mock, the value of the tracking tag becomes static and thus Dependabot PRs become useful.

See:
  - https://github.com/guardian/cdk/blob/b025f27d562c8897fa0b831612609a7e6b51b057/docs/001-general-usage.md#-mocking-gucdkversion