ci: Switch to a more direct upload to Riff-Raff
We've, somewhat randomly, started seeing builds fail when uploading artifacts.

The `riffRaffNotifyTeamcity` task is a proxy to TeamCity's `publishArtifacts` service message,
which allows:

> You can publish the build artifacts while the build is still running, immediately after the artifacts are built.

In this change, we switch to the `riffRaffUpload` task, which uploads artifacts at the _end_ of a build,
rather than as they are created.

The hope is this fixes the build, or provides more information to diagnose the underlying issue.

The sbt-riffraff-artifact plugin is also updated to 1.1.18.

In order to use the `riffRaffUpload` task, the `riffRaffUploadArtifactBucket` and `riffRaffUploadManifestBucket` properties need to be set.

Since https://github.com/guardian/sbt-riffraff-artifact/pull/63, the SBT plugin provides default values for these properties.

Lastly explcitly set the Riff-Raff project name.

When using `riffRaffNotifyTeamcity`, the project name defaults to the TeamCity project name (Editorial Tools::Typerigher).

When using `riffRaffUpload`, the project name defaults to the SBT project name (typerighter).

We need to override the default to remain consistent.

See:
  - https://github.com/guardian/typerighter/blob/05e3b077b192b2d9f05f84027095519a7fb30be4/script/citest#L71
  - https://www.jetbrains.com/help/teamcity/service-messages.html#Publishing+Artifacts+While+Build+is+in+Progress
  - https://github.com/guardian/sbt-riffraff-artifact/pull/63