Curses!
I went a bit deranged for a while because of a Null Pointer Exception. The debugger was useless, but eventually by writing tighter try catches around the db logic, it turned out we were actually dealing with a scalikejdbc.TooManyRowsException. Aha!

Turns out when it comes to dealing with multiple rows, using single() does not pluck a single item off (as I thought). Instead it throws the above exception - so a limit(1) operation is needed on our findLatestRevision method.

TL;DR - tighten your exceptions