DRY play-app commonalities
The commonLib project was not using javaOptions, because it's not deployable. DRYing the code removes the following warning:
```
[warn] there's a key that's not used by any other settings/tasks:
[warn]
[warn] * commonLib / Universal / javaOptions
[warn]   +- /Users/freddie_preece/Code/typerighter/build.sbt:33
[warn]
[warn] note: a setting might still be used by a command; to exclude a key from this `lintUnused` check
[warn] either append it to `Global / excludeLintKeys` or call .withRank(KeyRanks.Invisible) on the key
```

The new playProject abstraction changes the sub-project ID for rule manager from 'ruleManager' to 'rule-manager'. This brings consistency between the sub-project ID and its base folder name. In practice this means you would now need to execute `rule-manager / run` to run the rule manager sub-project.