Make sure rules with type dictionary are categorised as Collins Dictionary rules
To enable Rule Manager users to create dictionary rules that aren't part of Collins Dictionary. Previously we were taking using the category/source when creating the CheckerRuleResource. This will use the rule type instead, so we can use the Source to differentiate whether a dictionary rule comes from Collins or has been created by a user.