Only use Prout on endpoints without heavy caching
Of these two endpoints:

* https://manager.typerighter.gutools.co.uk/healthcheck - goes direct to ELB
* https://checker.typerighter.gutools.co.uk/healthcheck - is behind Cloudfront, and seems to have 1-hour cache time

...because of the long cache time on https://checker.typerighter.gutools.co.uk/healthcheck,
I think it's going to be too confusing to point Prout at it?

I think if Prout was making a POST request to checker.typerighter.gutools.co.uk, it would
presumably get a fresh response from the cache, but it would probably also need to
authenticate, for which in Prout there's no mechanism available at the moment.