Migrate to PRIVATE_WITH_EGRESS
PRIVATE_WITH_NAT is deprecated. Note that PRIVATE_WITH_EGRESS defaults to the same behaviour as PRIVATE_WITH_NAT, but has the option to set natGateways to 0.

If we are confident that our subnet will never need to make requests to the internet, we can opt to use PRIVATE_ISOLATED instead.