Add commit ID to Checker healthcheck to fix prout
As of [pull 465](https://github.com/guardian/typerighter/pull/465),
prout is configured to check both Rule Manager and Checker when pulls
are merged. However, that pull only added the commit ID to Rule Manager,
meaning every PR currently gets tagged by Prout with `Overdue on
Checker`. (Probably it wasn’t added because of the caching that used to
be on Checker, as described in [pull
466](https://github.com/guardian/typerighter/pull/466), though the prout
configuration probably should’ve been removed for that URL.)

Anyway, the caching was removed in [pull
468](https://github.com/guardian/typerighter/pull/468), so we can just
add the commit ID to the Checker healthcheck to fix prout: that’s what
this commit does.