Run CI on pull_request trigger, not push
When running on the `push` trigger, CI can provide misleading info in
the Github UI that makes it look like it has passed when in fact there
are failures. (See
https://github.com/guardian/mobile-apps-api/pull/2741#issuecomment-1777653733
and https://github.com/guardian/mobile-apps-api/pull/2760 for more
details.)

This commit updates the CI to instead run on `pull_request`, except for
the main branch.