Add dependency on jackson-module-scala 2.17.2
The [CI failed on this
PR](https://github.com/guardian/typerighter/actions/runs/16053987192/job/45303600828?pr=519)
after my latest change, with an error about jackson versions being
incompatible:

> com.fasterxml.jackson.databind.JsonMappingException: Scala module
> 2.14.3 requires Jackson Databind version >= 2.14.0 and < 2.15.0 - Found
> jackson-databind version 2.17.2

This is because our AWS dependencies don’t depend on
jackson-module-scala but do depend on jackson-databind, so we get their
newer version (2.17.2 currently) evicting the older version used by play
only for jackson-databind and not for jackson-module-scala.

To fix this, this commit adds a direct dependency on
jackson-module-scala at 2.17.2 (avoiding dependencyOverrides as [we’ve
identified some problems with using
it](https://github.com/guardian/maintaining-scala-projects/issues/20)).

One limitation of this is that if the AWS dependencies update their
dependency on jackson-databind to a newer version, we’ll get a new
version mismatch (indicated by a build failure on a PR updating the AWS
dependencies). To make it nice and obvious what’s gone wrong in these
cases, this commit also mentions the need to keep these in sync in a
comment.