Add task to check conflicting jackson versions
Recently, I tried to update the version of jackson used by this app in a
way that updated jackson-databind but didn’t update
jackson-module-scala. This caused a runtime check in the latter to fail
when the tests were run in CI. However, it’d be nice to be able to fail
earlier, since we always know those versions are incompatible.

This commit adds a new sbt task that errors if incomptaible versions of
jackson-databind and jackson-module-scala are used (i.e., ones that
differ in the two leading version digits).

The next step will be to get the compile task (or maybe the update
task?) to depend on this one so it’s executed automatically at
compilation.